refactor(launchdarkly-wrapper): rename client to adapter
