+++ packages/launchdarkly-adapter/modules/adapter.js
